added some css and bootstrap, still need to fix searching by power level

+++ src/main/java/com/sg/herosightings/dao/dto/Location.java
@@ -18,8 +18,8 @@ public class Location {
     private String localName;
     private String description;
     private String address;
-    private double latitude;
-    private double longitude;
+    private float latitude;
+    private float longitude;
     private Organization org;
     private List<Sighting> sightings = new ArrayList();
 
@@ -59,7 +59,7 @@ public class Location {
         return latitude;
     }
 
-    public void setLatitude(double latitude) {
+    public void setLatitude(float latitude) {
         this.latitude = latitude;
     }
 
@@ -67,7 +67,7 @@ public class Location {
         return longitude;
     }
 
-    public void setLongitude(double longitude) {
+    public void setLongitude(float longitude) {
         this.longitude = longitude;
     }
